scripts/requirements.txt
google-generativeai>=0.4.0
requests>=2.28.0
Pillow>=9.0.0
